requirements.txt
aiodns
aiofiles
aiohttp
aiohttp[speedups]
asyncpraw
cchardet
cryptg
gTTS
hachoir
heroku3
howdoi
pillow
python-dotenv
psutil
speedtest-cli
telethon
Pafy
youtube-dl